requirements.txt
streamlit
openai
python-dotenv
docx2txt
scikit-learn
pydub
chromadb
langchain_community
nbformat
fitz
frontend
pymupdf